# Optimize several linear algebra operations