This commit (which inherits largely from #901) implements a series of optimizations to the ways linear algebra operations are performed in the steppers and parameter transporters. These optimizations largely leverage the in-place operations provided by algebra plugins.