@stephenswat 2025-06-26
### Before

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   48564724 ns     48479844 ns           14 TracksPropagated=5.39783M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   15878392 ns     15849681 ns           44 TracksPropagated=16.5095M/s
BM_PROPAGATION_WIRE_CHAMBER_W_COV_TRANSPORT_262144_TRACKS/real_time   56845985 ns     56744657 ns           12 TracksPropagated=4.61148M/s
BM_PROPAGATION_WIRE_CHAMBER_262144_TRACKS/real_time                   17001492 ns     16970615 ns           41 TracksPropagated=15.4189M/s
```

### After

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   38872462 ns     38803657 ns           18 TracksPropagated=6.74369M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   16329409 ns     16298392 ns           43 TracksPropagated=16.0535M/s
BM_PROPAGATION_WIRE_CHAMBER_W_COV_TRANSPORT_262144_TRACKS/real_time   36978413 ns     36909221 ns           19 TracksPropagated=7.08911M/s
BM_PROPAGATION_WIRE_CHAMBER_262144_TRACKS/real_time                   13491872 ns     13467498 ns           52 TracksPropagated=19.4298M/s
```

Thus this yields a performance increase of up to 53% when covariance transport is enabled. Notably the performance of the toy detector benchmark without covariance transport degrades slightly.