@stephenswat 2025-06-26
God damn it I hate the SYCL compiler _so much_.